requirements.txt
aiogram==3.10.0
sqlalchemy[asyncio]==2.0.31
aiosqlite==0.20.0
asyncpg==0.29.0
jinja2==3.1.4
httpx==0.27.0
pdfplumber==0.11.2
pytesseract==0.3.10
pdf2image==1.17.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
